Start splitting common tests in multiple files
Extract tests using custom MyOptionalString type into a separate file.

This commit is best viewed using Git --color-moved option.
